move WaitUntilDisplayed() to IWebElementExtensions.cs for a more seamless usage
